changes made to notes
