+++ config/routes.rb
@@ -24,10 +24,12 @@ Rails.application.routes.draw do
         resources :person_project, shallow: true, only: :create
       end
       resources :person_project, only: %i[index show update destroy]
-      resources :projects, only: %i[create index show update destroy]
+      resources :projects, only: %i[create index show update destroy] do
+        scope module: :projects do
+          resources :notes, only: %i[create update destroy]
+        end
+      end
       resources :technologies, only: %i[index show]
-
-      resources :notes, only: %i[show create update destroy]
 
     end
   end
